backend/microfinance_project/requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
python-dateutil==2.9.0